test : 기존 테스트 통합
